+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.28](https://github.com/realityforge/proton/tree/v0.28) (2020-01-29) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.27...v0.28)
 
 Changes in this release: