+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `ElementsUtil.isElementDeprecated(Element)` helper method.
 * Add `ElementsUtil.isEffectivelyPublic(Element)` helper method.
+* Add `AbstractStandardProcessor.errorIfProcessingOverAndDeferredTypesUnprocessed()` helper method.
 
 ### [v0.28](https://github.com/realityforge/proton/tree/v0.28) (2020-01-29) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.27...v0.28)