+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.29](https://github.com/realityforge/proton/tree/v0.29) (2020-01-29) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.28...v0.29)
+
+Changes in this release:
 
 * Add `ElementsUtil.isElementDeprecated(Element)` helper method.
 * Add `ElementsUtil.isEffectivelyPublic(Element)` helper method.